Refactored the DecayRatesGraphView so the functionality for the decay rates tab is an extension of a more generic base class that will also be used to create the graphs in the Measurement and Dating Game tabs.
